reducing to smaller components
