+++ src/frontend/src/modals/IOModal/components/chatView/chatMessage/helpers/convert-files.ts
@@ -1,0 +1,20 @@
+export const convertFiles = (
+  files:
+    | (
+        | string
+        | {
+            path: string;
+            type: string;
+            name: string;
+          }
+      )[]
+    | undefined,
+) => {
+  if (!files) return [];
+  return files.map((file) => {
+    if (typeof file === "string") {
+      return file;
+    }
+    return file.path;
+  });
+};
