+++ src/frontend/src/modals/IOModal/newModal.tsx
@@ -8,7 +8,6 @@ import { useEffect, useState } from "react";
 import IconComponent from "../../components/common/genericIconComponent";
 import ShadTooltip from "../../components/common/shadTooltipComponent";
 import { Button } from "../../components/ui/button";
-import { InputOutput } from "../../constants/enums";
 import useAlertStore from "../../stores/alertStore";
 import useFlowStore from "../../stores/flowStore";
 import useFlowsManagerStore from "../../stores/flowsManagerStore";
@@ -16,10 +15,10 @@ import { useMessagesStore } from "../../stores/messagesStore";
 import { IOModalPropsType } from "../../types/components";
 import { cn } from "../../utils/utils";
 import BaseModal from "../baseModal";
-import IOFieldView from "./components/IOFieldView";
-import SessionSelector from "./components/IOFieldView/components/sessionSelector/newSessionSelector";
-import SessionView from "./components/SessionView";
 import ChatView from "./components/chatView/newChatView";
+import { ChatViewWrapper } from "./components/chatViewWrapper";
+import { SelectedViewField } from "./components/selectedViewField";
+import { SidebarOpenView } from "./components/sidebarOpenView";
 
 export default function IOModal({
   children,
@@ -129,7 +128,6 @@ export default function IOModal({
       ),
     ),
   );
-  const flowPool = useFlowStore((state) => state.flowPool);
   const [sessionId, setSessionId] = useState<string>(currentFlowId);
   const { isFetched: messagesFetched } = useGetMessagesQuery(
     {
@@ -280,217 +278,51 @@ export default function IOModal({
                 {sidebarOpen && <div className="font-semibold">Playground</div>}
               </div>
               {sidebarOpen && (
-                <div className="flex flex-col pl-3">
-                  <div className="flex flex-col gap-2 pb-2">
-                    <div className="flex items-center justify-between">
-                      <div className="flex items-center gap-2">
-                        <IconComponent
-                          name="MessagesSquare"
-                          className="h-[18px] w-[18px] text-ring"
-                        />
-                        <div className="text-[13px] font-normal">Chat</div>
-                      </div>
-                      <ShadTooltip styleClasses="z-50" content="New Chat">
-                        <div>
-                          <Button
-                            data-testid="new-chat"
-                            variant="ghost"
-                            className="flex h-8 w-8 items-center justify-center !p-0 hover:bg-secondary-hover"
-                            onClick={(_) => {
-                              setvisibleSession(undefined);
-                              setSelectedViewField(undefined);
-                            }}
-                          >
-                            <IconComponent
-                              name="Plus"
-                              className="h-[18px] w-[18px] text-ring"
-                            />
-                          </Button>
-                        </div>
-                      </ShadTooltip>
-                    </div>
-                  </div>
-                  <div className="flex flex-col">
-                    {sessions.map((session, index) => (
-                      <SessionSelector
-                        setSelectedView={setSelectedViewField}
-                        selectedView={selectedViewField}
-                        key={index}
-                        session={session}
-                        deleteSession={(session) => {
-                          handleDeleteSession(session);
-                          if (selectedViewField?.id === session) {
-                            setSelectedViewField(undefined);
-                          }
-                        }}
-                        updateVisibleSession={(session) => {
-                          setvisibleSession(session);
-                        }}
-                        toggleVisibility={() => {
-                          setvisibleSession(session);
-                        }}
-                        isVisible={visibleSession === session}
-                        inspectSession={(session) => {
-                          setSelectedViewField({
-                            id: session,
-                            type: "Session",
-                          });
-                        }}
-                      />
-                    ))}
-                  </div>
-                </div>
+                <SidebarOpenView
+                  sessions={sessions}
+                  setSelectedViewField={setSelectedViewField}
+                  setvisibleSession={setvisibleSession}
+                  handleDeleteSession={handleDeleteSession}
+                  visibleSession={visibleSession}
+                  selectedViewField={selectedViewField}
+                />
               )}
             </div>
           </div>
           <div className="flex h-full min-w-96 flex-grow bg-background">
             {selectedViewField && (
-              <div
-                className={cn(
-                  "flex h-full w-full flex-col items-start gap-4 p-4",
-                  !selectedViewField ? "hidden" : "",
-                )}
-              >
-                <div className="font-xl flex items-center justify-center gap-3 font-semibold">
-                  {haveChat && (
-                    <button onClick={() => setSelectedViewField(undefined)}>
-                      <IconComponent
-                        name={"ArrowLeft"}
-                        className="h-6 w-6"
-                      ></IconComponent>
-                    </button>
-                  )}
-                  {
-                    nodes.find((node) => node.id === selectedViewField.id)?.data
-                      .node.display_name
-                  }
-                </div>
-                <div className="h-full w-full">
-                  {inputs.some(
-                    (input) => input.id === selectedViewField.id,
-                  ) && (
-                    <IOFieldView
-                      type={InputOutput.INPUT}
-                      left={false}
-                      fieldType={selectedViewField.type!}
-                      fieldId={selectedViewField.id!}
-                    />
-                  )}
-                  {outputs.some(
-                    (output) => output.id === selectedViewField.id,
-                  ) && (
-                    <IOFieldView
-                      type={InputOutput.OUTPUT}
-                      left={false}
-                      fieldType={selectedViewField.type!}
-                      fieldId={selectedViewField.id!}
-                    />
-                  )}
-                  {sessions.some(
-                    (session) => session === selectedViewField.id,
-                  ) && (
-                    <SessionView
-                      session={selectedViewField.id}
-                      id={currentFlowId}
-                    />
-                  )}
-                </div>
-              </div>
+              <SelectedViewField
+                selectedViewField={selectedViewField}
+                setSelectedViewField={setSelectedViewField}
+                haveChat={haveChat}
+                inputs={inputs}
+                outputs={outputs}
+                sessions={sessions}
+                currentFlowId={currentFlowId}
+                nodes={nodes}
+              />
             )}
-            <div
-              className={cn(
-                "flex h-full w-full flex-col justify-between p-4",
-                selectedViewField ? "hidden" : "",
-              )}
-            >
-              <div className="mb-4 h-[5%] text-[16px] font-semibold">
-                {visibleSession && sessions.length > 0 && sidebarOpen && (
-                  <div className="hidden lg:block">
-                    {visibleSession === currentFlowId
-                      ? "Default Session"
-                      : `${visibleSession}`}
-                  </div>
-                )}
-                <div className={cn(sidebarOpen ? "lg:hidden" : "")}>
-                  <div className="flex items-center gap-2">
-                    <Button
-                      variant="ghost"
-                      size="icon"
-                      onClick={() => setSidebarOpen(true)}
-                      className="h-8 w-8"
-                    >
-                      <IconComponent
-                        name={"PanelLeftOpen"}
-                        className="h-[18px] w-[18px] text-ring"
-                      />
-                    </Button>
-                    <div className="font-semibold">Playground</div>
-                  </div>
-                </div>
-                <div
-                  className={cn(
-                    sidebarOpen ? "pointer-events-none opacity-0" : "",
-                    "absolute flex h-8 items-center justify-center rounded-sm ring-offset-background transition-opacity focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
-                    isPlayground ? "right-2 top-4" : "right-12 top-2",
-                  )}
-                >
-                  <ShadTooltip
-                    side="bottom"
-                    styleClasses="z-50"
-                    content="New Chat"
-                  >
-                    <Button
-                      className="mr-2 h-[32px] w-[32px] hover:bg-secondary-hover"
-                      variant="ghost"
-                      size="icon"
-                      onClick={(_) => {
-                        setvisibleSession(undefined);
-                        setSelectedViewField(undefined);
-                      }}
-                    >
-                      <IconComponent
-                        name="Plus"
-                        className="!h-[18px] !w-[18px] text-ring"
-                      />
-                    </Button>
-                  </ShadTooltip>
-                  {!isPlayground && <Separator orientation="vertical" />}
-                </div>
-              </div>
-              {haveChat ? (
-                <div
-                  className={cn(
-                    visibleSession ? "h-[95%]" : "h-full",
-                    sidebarOpen
-                      ? "pointer-events-none blur-sm lg:pointer-events-auto lg:blur-0"
-                      : "",
-                  )}
-                >
-                  {messagesFetched && (
-                    <ChatView
-                      focusChat={sessionId}
-                      sendMessage={sendMessage}
-                      chatValue={chatValue}
-                      setChatValue={setChatValue}
-                      lockChat={lockChat}
-                      setLockChat={setLockChat}
-                      visibleSession={visibleSession}
-                      closeChat={
-                        !canvasOpen
-                          ? undefined
-                          : () => {
-                              setOpen(false);
-                            }
-                      }
-                    />
-                  )}
-                </div>
-              ) : (
-                <span className="flex h-full w-full items-center justify-center font-thin text-muted-foreground">
-                  Select an IO component to view
-                </span>
-              )}
-            </div>
+            <ChatViewWrapper
+              selectedViewField={selectedViewField}
+              visibleSession={visibleSession}
+              sessions={sessions}
+              sidebarOpen={sidebarOpen}
+              currentFlowId={currentFlowId}
+              setSidebarOpen={setSidebarOpen}
+              isPlayground={isPlayground}
+              setvisibleSession={setvisibleSession}
+              setSelectedViewField={setSelectedViewField}
+              haveChat={haveChat}
+              messagesFetched={messagesFetched}
+              sessionId={sessionId}
+              sendMessage={sendMessage}
+              chatValue={chatValue}
+              setChatValue={setChatValue}
+              lockChat={lockChat}
+              setLockChat={setLockChat}
+              canvasOpen={canvasOpen}
+              setOpen={setOpen}
+            />
           </div>
         </div>
       </BaseModal.Content>
