+++ src/frontend/src/modals/IOModal/types/chat-view-wrapper.ts
@@ -1,0 +1,27 @@
+import { ButtonProps } from "@/components/ui/button";
+import { IconComponentProps, TooltipProps } from "@/types/components";
+import { FC } from "react";
+
+export type ChatViewWrapperProps = {
+  selectedViewField: { type: string; id: string } | undefined;
+  visibleSession: string | undefined;
+  sessions: string[];
+  sidebarOpen: boolean;
+  currentFlowId: string;
+  setSidebarOpen: (open: boolean) => void;
+  isPlayground: boolean | undefined;
+  setvisibleSession: (session: string | undefined) => void;
+  setSelectedViewField: (
+    field: { type: string; id: string } | undefined,
+  ) => void;
+  haveChat: { type: string; id: string; displayName: string } | undefined;
+  messagesFetched: boolean;
+  sessionId: string;
+  sendMessage: (options: { repeat: number; files?: string[] }) => Promise<void>;
+  chatValue: string;
+  setChatValue: (value: string) => void;
+  lockChat: boolean;
+  setLockChat: (locked: boolean) => void;
+  canvasOpen: boolean | undefined;
+  setOpen: (open: boolean) => void;
+};
