+++ src/frontend/src/modals/IOModal/types/sidebar-open-view.ts
@@ -1,0 +1,10 @@
+export type SidebarOpenViewProps = {
+  sessions: string[];
+  setSelectedViewField: (
+    field: { type: string; id: string } | undefined,
+  ) => void;
+  setvisibleSession: (session: string | undefined) => void;
+  handleDeleteSession: (session: string) => void;
+  visibleSession: string | undefined;
+  selectedViewField: { type: string; id: string } | undefined;
+};
